requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.32
pydantic==2.8.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
